src_docker/smi/requirements.txt
google-api-core
google-auth
google-auth-oauthlib
google-cloud-bigquery-storage
google-cloud-core
google-crc32c
google-resumable-media
googleapis-common-protos
google-api-python-client
googleads
pandas
pandas-gbq
boto3
pytz
pydantic
pydantic[email]
google-cloud-bigquery
elementary-data[bigquery]
oauth2client
dateparser
flake8
s3fs
tenacity
botocore
dateparser
pyarrow
